Mac ARM64 integration (#19571)
* Update Mac platform to support multi-arch detection
* New Mac arm64 specific cmake files
* Remove EditorCommon.Headers causing issues with Qt6 moving and Mac

* Fix QAbstractNativeEventFilter::nativeEventFilter Signature
* Update deprecated constructor call to QMouseEvent
* Update deprecated QHoverEvent->pos() to ->position)
* Fix issue where file that needs mocing 'QtEditorApplication_mac.h' was not in any cmake file list, thus not being auto-moc'd
* Removed unneeded include for that moc
* Added static_cast<int> in numerous places to fix error "Implicit conversion loses integer precision: 'qsizetype' (aka 'long long') to 'int'"



Unit test fixes and review feedback

various compilation fixes


Mac support

+++ Code/Editor/AssetImporter/AssetImporterManager/AssetImporterDragAndDropHandler.cpp
@@ -104,7 +104,7 @@ void AssetImporterDragAndDropHandler::ProcessDragEnter(QDragEnterEvent* event)
     }
 
     QList<QUrl> urlList = mimeData->urls();
-    int urlListSize = urlList.size();
+    int urlListSize = static_cast<int>(urlList.size());
 
     // runs through the file list first and checks for any "crate" files - if it finds ANY, return (and don't accept the event)
     for (int i = 0; i < urlListSize; ++i)
