+++ Code/Framework/AzQtComponents/AzQtComponents/Components/FilteredSearchWidget.cpp
@@ -370,7 +370,7 @@ namespace AzQtComponents
         QMap<QString, QList<QStandardItem*>> categoryToEntryItems; // category name to sub-items that will be added
         
 
-        const int numItems = unfilteredData.length();
+        const int numItems = static_cast<int>(unfilteredData.length());
         for (int unfilteredDataIndex = 0; unfilteredDataIndex < numItems; ++unfilteredDataIndex)
         {
             const SearchTypeFilter& filter = unfilteredData.at(unfilteredDataIndex);
@@ -395,7 +395,7 @@ namespace AzQtComponents
             }
         }
 
-        const int numCategories = categoriesInOrder.size();
+        const int numCategories = static_cast<int>(categoriesInOrder.size());
 
         // If there is only one category and its name is empty, discard it,
         // and add its children directly to the model as one big column of N rows
@@ -895,7 +895,7 @@ namespace AzQtComponents
         if (!typeFilter.displayName.isEmpty())
         {
             m_typeFilters.append(typeFilter);
-            SetFilterStateByIndex(m_typeFilters.length() - 1, typeFilter.enabled);
+            SetFilterStateByIndex(static_cast<int>(m_typeFilters.length()) - 1, typeFilter.enabled);
         }
 
         SetTypeFilterVisible(true);
@@ -961,7 +961,7 @@ namespace AzQtComponents
 
     int FilteredSearchWidget::GetTypeFilterCount()
     {
-        return m_typeFilters.size();
+        return static_cast<int>(m_typeFilters.size());
     }
 
     void FilteredSearchWidget::GetTypeFilterDetails(const int index, QString& categoryKeyOut, QString& displayNameOut, bool& enabledOut)
@@ -1308,7 +1308,7 @@ namespace AzQtComponents
         ConfigHelpers::GroupGuard guard(&settings, widgetName);
         settings.setValue(g_textFilterKey, textFilter());
 
-        const int size = m_typeFilters.size();
+        const int size = static_cast<int>(m_typeFilters.size());
         settings.beginWriteArray(g_typeFiltersKey, size);
         for (int i = 0; i < size; ++i)
         {
@@ -1553,7 +1553,7 @@ namespace AzQtComponents
                 do
                 {
                     // Find filter term within the text.
-                    highlightTextIndex = label.lastIndexOf(m_selector->GetFilterString(), highlightTextIndex - 1, Qt::CaseInsensitive);
+                    highlightTextIndex = static_cast<int>(label.lastIndexOf(m_selector->GetFilterString(), highlightTextIndex - 1, Qt::CaseInsensitive));
                     if (highlightTextIndex >= 0)
                     {
                         // Insert background-color terminator at appropriate place to return to normal text.
