+++ Code/Framework/AzQtComponents/AzQtComponents/Components/StyledLineEdit.cpp
@@ -53,7 +53,7 @@ namespace AzQtComponents
     void StyledLineEdit::validateEntry()
     {
         QString textToValidate = text();
-        int length = textToValidate.length();
+        int length = static_cast<int>(textToValidate.length());
 
         if (!validator())
         {
